backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
nltk>=3.8.0
pydantic>=2.0.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
requests>=2.31.0
Pillow>=10.0.0
pdfplumber>=0.10.0
python-magic>=0.4.27
python-dotenv==1.0.0
# Note: Database dependencies removed for Windows compatibility
# You can add them back later when needed 